refactor(api): restructure knowledge services and enhance document/resource handling
- Replaced KnowledgeService with ResourceService and DocumentService in various modules for improved separation of concerns.
- Updated CanvasService and KnowledgeController to utilize the new services for document and resource operations.
- Ensured adherence to Tailwind CSS styling conventions and optimized property access using optional chaining and nullish coalescing.
- Removed deprecated code and improved overall code maintainability.
