+++ apps/api/src/modules/knowledge/knowledge.processor.ts
@@ -2,7 +2,6 @@ import { Processor, WorkerHost } from '@nestjs/bullmq';
 import { Logger } from '@nestjs/common';
 import { Job } from 'bullmq';
 
-import { KnowledgeService } from './knowledge.service';
 import {
   QUEUE_DELETE_KNOWLEDGE_ENTITY,
   QUEUE_POST_DELETE_KNOWLEDGE_ENTITY,
@@ -13,12 +12,14 @@ import {
   FinalizeResourceParam,
   PostDeleteKnowledgeEntityJobData,
 } from './knowledge.dto';
+import { ResourceService } from './resource.service';
+import { DocumentService } from './document.service';
 
 @Processor(QUEUE_RESOURCE)
 export class ResourceProcessor extends WorkerHost {
   private readonly logger = new Logger(ResourceProcessor.name);
 
-  constructor(private knowledgeService: KnowledgeService) {
+  constructor(private resourceService: ResourceService) {
     super();
   }
 
@@ -26,7 +27,7 @@ export class ResourceProcessor extends WorkerHost {
     this.logger.log(`[${QUEUE_RESOURCE}] job: ${JSON.stringify(job)}`);
 
     try {
-      await this.knowledgeService.finalizeResource(job.data);
+      await this.resourceService.finalizeResource(job.data);
     } catch (error) {
       this.logger.error(`[${QUEUE_RESOURCE}] error: ${error?.stack}`);
       throw error;
@@ -38,7 +39,10 @@ export class ResourceProcessor extends WorkerHost {
 export class DeleteKnowledgeEntityProcessor extends WorkerHost {
   private readonly logger = new Logger(DeleteKnowledgeEntityProcessor.name);
 
-  constructor(private knowledgeService: KnowledgeService) {
+  constructor(
+    private resourceService: ResourceService,
+    private documentService: DocumentService,
+  ) {
     super();
   }
 
@@ -49,9 +53,9 @@ export class DeleteKnowledgeEntityProcessor extends WorkerHost {
 
     try {
       if (entityType === 'resource') {
-        await this.knowledgeService.deleteResource({ uid }, entityId);
+        await this.resourceService.deleteResource({ uid }, entityId);
       } else if (entityType === 'document') {
-        await this.knowledgeService.deleteDocument({ uid }, { docId: entityId });
+        await this.documentService.deleteDocument({ uid }, { docId: entityId });
       }
     } catch (error) {
       this.logger.error(`[${QUEUE_DELETE_KNOWLEDGE_ENTITY}] error: ${error?.stack}`);
@@ -64,7 +68,10 @@ export class DeleteKnowledgeEntityProcessor extends WorkerHost {
 export class PostDeleteKnowledgeEntityProcessor extends WorkerHost {
   private readonly logger = new Logger(PostDeleteKnowledgeEntityProcessor.name);
 
-  constructor(private knowledgeService: KnowledgeService) {
+  constructor(
+    private resourceService: ResourceService,
+    private documentService: DocumentService,
+  ) {
     super();
   }
 
@@ -75,9 +82,9 @@ export class PostDeleteKnowledgeEntityProcessor extends WorkerHost {
 
     try {
       if (entityType === 'resource') {
-        await this.knowledgeService.postDeleteResource({ uid }, entityId);
+        await this.resourceService.postDeleteResource({ uid }, entityId);
       } else if (entityType === 'document') {
-        await this.knowledgeService.postDeleteDocument({ uid }, entityId);
+        await this.documentService.postDeleteDocument({ uid }, entityId);
       }
     } catch (error) {
       this.logger.error(`[${QUEUE_POST_DELETE_KNOWLEDGE_ENTITY}] error: ${error?.stack}`);
