+++ apps/api/src/modules/tool/internal-tool.service.ts
@@ -3,7 +3,7 @@ import { throttle } from 'lodash';
 import { User, UpsertDocumentRequest, CodeArtifactType } from '@refly/openapi-schema';
 import { BaseChatModel } from '@langchain/core/language_models/chat_models';
 import { PrismaService } from '../common/prisma.service';
-import { KnowledgeService } from '../knowledge/knowledge.service';
+import { DocumentService } from '../knowledge/document.service';
 import { CollabService } from '../collab/collab.service';
 import { CollabContext } from '../collab/collab.dto';
 import { ProviderService } from '../provider/provider.service';
@@ -24,7 +24,7 @@ export class InternalToolService {
 
   constructor(
     private readonly prisma: PrismaService,
-    private readonly knowledgeService: KnowledgeService,
+    private readonly documentService: DocumentService,
     private readonly providerService: ProviderService,
     private readonly collabService: CollabService,
     private readonly canvasSyncService: CanvasSyncService,
@@ -73,7 +73,7 @@ export class InternalToolService {
         canvasId,
       };
 
-      const document = await this.knowledgeService.createDocument(user, documentRequest);
+      const document = await this.documentService.createDocument(user, documentRequest);
 
       // Add node to canvas if canvasId is provided
       if (canvasId) {
